파일 구조 정리 및 개발 밑작업
- NPCAIController 및 Base 생성
  - 관련 코드 정리
  - Dynamic, Static 컨트롤러 삭제 및 Base로 통합

ToDo
- 플레이어 상호작용 오브젝트 아웃라인 하이라이트 기능 제작
- Mob 클래스 및 태스크 제작
