+++ Content/Characters/NPC/BP_MovePoint.uasset
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:816a12e87c244b2fd01a041035f5e2bf8283b7695a770f1fc9db8c321be4e0f9
+size 22725
