--- Content/Characters/NPCBase/Blueprint/AIController/BP_DynamicNPCAIController.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:6ed0b61443d6647d8c70df2c18c55e1374c26a56dba2645f6a05bbd59718a26e
-size 20810
