--- Content/Characters/NPCBase/Blueprint/AIController/BP_NPCAIController.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:55ac0fcb8faa7b6fc1510e63eecd435137ae820539ea63ed87e871a4a10f7da6
-size 20586
