--- Content/Characters/NPCBase/Blueprint/AIController/BP_StaticNPCAIController.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:3c39bc3ab35034c2f5f9612eb26aca2c745b6c1aa5f1174f03fe72fde8efa2ae
-size 20778
