--- Content/Characters/NPCBase/Blueprint/BP_DynamicNPC.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:f8e5efd58e454f1352bd14b61338f48cedf67e402a20024fa73d7bf15860a580
-size 44472
