+++ Content/Characters/NPCBase/Blueprint/BP_MovePoint.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:c3300abb4d8e81d1100180e1c2cf0b440a9d8d14fd4067d09721e5f026e94341
-size 21658
+oid sha256:d2c1724f1cf96cb8359a3c5132cf0b3b1feca68107b03117878df71bb760da30
+size 2393
