+++ Content/Characters/NPCBase/Blueprint/BP_NPCBase.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:bf366b880c90539990fb1af07dab69fb25508340e6694c75bab5b7535fe3dfcc
-size 36873
+oid sha256:157453cea7769fe239dc8aafdc4d30035502375e42a81b0435b9f1da3811f555
+size 36478
