--- Content/Characters/NPCBase/Blueprint/BP_StaticNPC.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:2346a96f480a0e812d72b0f41dc849e087c39cc4292aaf9976f0100f2b2b53e5
-size 43609
