+++ Source/Dungeon_Armory/Private/AI/Tasks/BTTask_ChangeSpeed.cpp
@@ -5,8 +5,8 @@
 
 #include "BehaviorTree/BlackboardComponent.h"
 
+#include "Characters/NPC/NPCAIControllerBase.h"
 #include "Characters/NPC/NPCBase.h"
-#include "Characters/NPC/NPCAIController.h"
 
 UBTTask_ChangeSpeed::UBTTask_ChangeSpeed()
 {
@@ -15,7 +15,7 @@ UBTTask_ChangeSpeed::UBTTask_ChangeSpeed()
 
 EBTNodeResult::Type UBTTask_ChangeSpeed::ExecuteTask(UBehaviorTreeComponent& OwnerComp, uint8* NodeMemory)
 {
-	auto NPCController = Cast<ANPCAIController>(OwnerComp.GetAIOwner());
+	auto NPCController = Cast<ANPCAIControllerBase>(OwnerComp.GetAIOwner());
 	if (!NPCController)
 	{
 		return EBTNodeResult::Failed;
