+++ Source/Dungeon_Armory/Private/AI/Tasks/BTTask_Stay.cpp
@@ -5,7 +5,7 @@
 #include "AIController.h"
 #include "BehaviorTree/BehaviorTreeTypes.h"
 #include "BehaviorTree/BlackboardComponent.h"
-#include "Characters/NPC/NPCAIController.h"
+#include "Characters/NPC/NPCAIControllerBase.h"
 #include "Manager/InGameTimeManager.h"
 
 UBTTask_Stay::UBTTask_Stay()
