+++ Source/Dungeon_Armory/Private/Characters/Mob/MobBase.cpp
@@ -1,0 +1,5 @@
+// Fill out your copyright notice in the Description page of Project Settings.
+
+
+#include "Characters/Mob/MobBase.h"
+
