+++ Source/Dungeon_Armory/Public/Characters/Mob/MobBase.h
@@ -1,0 +1,17 @@
+// Fill out your copyright notice in the Description page of Project Settings.
+
+#pragma once
+
+#include "CoreMinimal.h"
+#include "Characters/NPC/NPCBase.h"
+#include "MobBase.generated.h"
+
+/**
+ * 
+ */
+UCLASS()
+class DUNGEON_ARMORY_API AMobBase : public ANPCBase
+{
+	GENERATED_BODY()
+	
+};
