--- Source/Dungeon_Armory/Public/Characters/NPC/DynamicNPCAIController.h
@@ -1,28 +1,0 @@
-// Fill out your copyright notice in the Description page of Project Settings.
-
-#pragma once
-
-#include "CoreMinimal.h"
-#include "Characters/NPC/NPCAIController.h"
-#include "DynamicNPCAIController.generated.h"
-
-UCLASS()
-class DUNGEON_ARMORY_API ADynamicNPCAIController : public ANPCAIController
-{
-	GENERATED_BODY()
-
-/***** Functions (Unreal) *****/
-
-protected:
-	// Called when the game starts or when spawned
-	virtual void BeginPlay() override;
-	// Called when the controller possess a Pawn
-	virtual void OnPossess(APawn* InPawn) override;
-
-/***** Functions (Behavior Tree) *****/
-private:
-	UFUNCTION()
-	void OnStayCompleted();
-	UFUNCTION()
-	void OnRoamToPointReached(FAIRequestID RequestID, EPathFollowingResult::Type Result);
-};
